requirements.txt
ansible
Jinja2
netaddr
PyYAML
ncclient
lxml
junos-eznc
jxmlease
ruamel.yaml
colorama
pyfiglet
argcomplete
